North/south road segments now line up correctly with intersections
When a road segment is at an angle (i.e. it's going up or down
a hill), then the length of the segment between the two
corresponding road intersections will differ depending on the
angle the segment rises/falls. Calculating the correct road
segment length is necessary to make sure it lines up with
the corresponding road intersections without gaps.

To calculate the length of the road segment, the 'distance
between two points' formula is used to calculate the distance
between the two intersections (and therefore, how long the
road segment should be).

I.e., sqrt((x2 - x1)^2 + (y2 - y1)^2)