Simplifying road/building generation to always follow terrain
I.e. road and building generation will no longer to avoid negative
y values - if the terrain has negative heights, then roads and
buildings will follow that. Instead, if we want to avoid negative
terrain heights, it is the responsibility of terrain generation
to handle that. This allows us to consolidate that logic in a
single place.