Lot size now determines max building height
The main motivator is to prevent super tall, skinny buildings
in the core, which I think looks better aesthetically and
seems more "realistic" (recent trends in New York City be
damned).

Since this results in fewer taller buildings in the core,
also reducing the exponent on height decay from the core
to compensate.