Fixing camera rotation, while also simplifying logic
I.e., when the camera rotates as a result of a new target
point being selected, it will now always do a 90deg turn
(instead of sometimes a 270deg turn), and the camera will
also turn to face the new movement direction, instead of
sometimes facing some different direction.

Also, I think the new logic to determine the new rotation
angle is easier to understand.