Adding experimental birds eye view animiation
The camera suddenly rises quickly to a high height, and looks
straight down at the city below. Sort of like you're viewing
the city from space.

This doesn't handle rotations when changing x/z direction to
move to a new point. These rotations always happen around the
y-axis, and since the x-axis is rotated, this causes the
y-axis to be rotated as well. The end result is that rotations
around the y-axis don't rotate in the direction they are
intended to go.