Initial implementation of road network
Stores the road network as a collection of intersections,
with edges to other intersections.

This is now used for determining which roads to render,
and in the future can be used for controlling the camera
(i.e. so that it only follows roads, and not random
terrain).

Currently, as a first step, the road network mimics the
old "assume there are roads everywhere" structure, so this
shouldn't result in any visible changes yet.