Adding road network implementation with 'additive' algorithm
I.e. this implementation starts at the center of the map, and
builds the road network out from there. This is the opposite
approach of the previous implementation, which builds a fully
populated rectangular network around the whole city, and then
removes edges.

The benefit of this new implementation is that it makes it easier
to build a more irregular network in which all of the roads
connect to each other, and which also more automatically follows
the terrain (for example, filling in a valley but not expanding
out of it unless the terrain is flat enough for roads to make it
out. With the 'subtractive' algorithm, it's a lot more complicated
to do this while making sure the roads are all connected to
each other.

Both implementations have the same interface and so can be
swapped out for each other.