Decoupling motion of camera from rotation of camera
Doing this by attaching the camera to a separate object. The
object is then moved around (automatically moving the camera
along with it), while the camera is free to rotate and look
at arbitrary angles without messing up the rotations of the
base object.

Conceptually this is like attaching the camera to a pole,
walking around holding pole. The camera will get moved
around the scene, and can rotate without affecting the
movement of the pole.

Besides generally making camera movement more flexible, in
the immediate term this allows the bird's eye view to work
correctly, by preventing the camera from rolling onto its
side.