Adding path finder that navigates to arbitrary point in road network
Previously existing implementation can only navigate to a point
-180 or 180 degrees from the starting point, and is not aware
of the road network and so will happily go over open terrain.

Implementation uses Dijktra's algorithm to implement the path
finding. On my local computer it takes 10-30ms to find the route
to the next target point, so will need to see if this causes
noticeable pauses in the animation when a new target point is
chosen.