Fixing logic to determine x/z movement amount
Previously it assumed that x and z motion would alternate.
That is, the camera would move along the x axis for a certain
amount to the next target point, then along the z axis to the
next target point, then the x axis again to the next target
point, etc. However, with the new Dijkstra's algorithm
path finder multiple target points along the same axis
will be followed.

Instead, this implementation now compares the location of the
target point from the current point, and determines the
x/z motion as needed. This still assumes that movement only
occurs on one axis at a time (i.e. there is no diagonal
movement).