Not performance has-road check if we know we won't build a building
I.e., breaking up the if check so that we short circuit if the
random probability of building on the lot is not met, since
determining whether to randomly build a building is faster
than checking to see if the lot has an adjacent road.

In Safari on my local computer, this seems to speed up
total building generation time by ~40-50ms.