Combining horizontal/vertical motion controllers into 'vehicle' controller
Models a vehicle which can drive on the roads, or fly into the air.
Since horizontal and vertical parameters can be coupled to each other
in practice (e.g. whether or not the path finder is required to follow
roads or not can depend on whether or not the vehicle is on the ground
or in the air), so making this coupling explicit.