Preventing vehicle from going below the road/terrain
Previously the vehicle could go below the road, but the camera
was forced up to the terrain height. Having the vehicle respect
the terrain height is cleaner and also fixes an issue with the
zoom to bird's eye animation in which it would look at the
ground too early in the ascent. (Due to the 'look down'
animation happening while the vehicle was spending extra time
moving from below ground to the surface).