Allowing camera to move based on a sine wave
I.e. instead of moving linearly toward a target position,
either starting fast and slowing down to reach the target,
or starting slow and speeding up as the target is reached.

Immediate motivation is to improve the debug animation so that
the edges of the world aren't shown when zooming up/down to the
'from space' view, by controlling the rate of the camera's
x rotation as it ascends/descends.