Preventing camera jump when debug 'space view' zoom down completes
Was caused previously by the camera navigating to the place the
vehicle was when the descent started, instead of the final place
the vehicle would be when the descent was complete. As a result,
the camera would suddenly jump from the end point of the descent
to the current vehicle position.