Prevent buildings that don't fully go into ground on steep terrain
The problem is that the calculations of the terrain heights of a
building lot don't match up with how terrain is visually displayed,
and that discrepency can cause the building's foundation to be
above the corresponding point in the terrain triangle when
rendered.

As a workaround, the building's foundation is set to be the
lowest point in the bigger zoned block. Since terrain triangles
correspond to block boundaries, this should cause the terrain
height calculations to match up with what is displayed visually.