Adding canonical constants to remove repeated magic values
Simulates a static property on the Terrain "class", a la Java.
Conceptually, this seems like a good home, since these constants
aren't specific to instances, and there are cases where these
constants need to be used but a Terrain instance isn't present.
Time will tell if this is a good idea, but it seems like an
improvement.