First proto version of bridges
Each edge in the road network now has an attribute to
indicate if it follows the terrain, or is a bridge (i.e.
it goes above the terrain, instead of on top of it).

Bridges are not rendered as actual bridges yet, they
look like roads follow the height of water tiles.

There are too many bridges, there are not enough
constraints on them yet so it's not uncommen to have
like 8 bridges side by side. But, a max bridge
length can be configured, and perpendicular bridges
that run into each other are prevented.

Finally, building generation is smart enough to
not place buildings next to bridges (since it means
the building would be in water).