Bridges partially rendered as a bridge, not 'road on water'
I.e., the bridge road surface hovers in the air over water (or
land), instead of following the height of the underlaying
terrain.

Water is correctly rendered under the bridge, and the bridge
is displayed as a continuous surface. However, the height
of the bridge is currently hard-coded to an arbitrary value
(in this commit, "2"). Properly setting the bridge height
based on the terminal land road intersections will happen
in a future commit.