Road intersections have distinct height, separate from terrain height
Needed for bridges, so that the road deck on a bridge can have a
height that is determined by the height of its terminal road
intersections on the land. Currently, the road deck height is
still hard-coded to an arbitrary value for all bridges, but this
hard-coded value is read from the intersection itself, instead of
being hard-coded into the scene generation logic. In a future
commit, road intersection will be given "real" (i.e. non
hard-coded) height values.