Camera correctly follows bridge road deck
Previously, when in 'driving' mode the camera always based
its height on the underlaying terrain. Previously, the road
surface height and the terrain height were always the same,
so this was OK. With bridges, the road deck height is _not_
the same as the terrain height, with the result that the
camera would sink through the bridge and down to water height,
and then rise back up to land height on the other side.

The camera will now check to see if it is over a road, and
set the road height as its minimum height, falling back to
the terrain height if not currently over a road. The end
result is that the camera now properly 'drives' over bridges.