Allow dragging with mouse to pan around the city
I.e. the same way you can drag on the map in Google/Apple Maps
to pan around.

If you see something on the screen, for example a building,
it will move in the same direction that you move the mouse.

If you move the mouse to the right, the building will move
to the right, if you move it to the left if will move to
the left.

Depending on the camera tilt (i.e. rotation around X axis)
moving the mouse up will cause the building to either move
"up" toward the top of the screen (if looking straight down),
or farther away (if looking at an angle). Moving the mouse
down will cause the building to either move "down" toward
the bottom of the screen, or closer to the viewer.

These movements are relative to how to scene currently appears
on the screen, so the current rotation around the Y-axis has
to be taken into account. What we want is that moving the mouse
left/right always causes the scene _as currently displayed_
to move left/right. If the current Y rotation is not taken
into account, moving the mouse left/right will sometimes cause
the scene to move left/right, and sometimes up/down, depending
on the current Y rotation. (And a similar problem would exist
when moving the mouse up/down).

The current implementation requires setting hidden <input>
fields as hacky temporary way to communicate between the
"track the drag coordinates" logic and the "update camera
position" logic.