Experiment at eroding river bank to prevent steep cliffs
If there is a large variation between the highest terrai
next to a river bank and the lowest part, a steep canyon is
cut into the terrain when adding the river. This is
interesting sometimes, but if it happens all the time it's
not super realistic and is limiting.

The code added in this PR is a first attempt at having the
land near the river banks gradually recede in height instead
of always resulting in a steep cliff. This is to simulate the
river eroding the land over time, resulting in a V-shaped
river valley.

In this iteration, there are clear vertical artifacts in the
terrain, that have the appearance of many narrow north/south
canyons where the river bank was "eroded". That will need to
be fixed before this is ready for primetime, but this is a
first step.

Also, in this PR only the northern river bank is eroded.