All touches available in touch->camera syncing logic
Previously, only a single touch was available for determining
things like the camera's position, rotation, and zoom based
on the current mouse/touch coordinates. Now all touches are
available, which is an initial step toward implementing features
such as "pinch to zoom".

However, this commit should not cause any functional changes,
since the internal logic only uses the first touch in the list
(to implement "drag finger to pan").