Initial implementation of multi-touch camera tilt rotation
The multi-touch rotation support now has two modes. If the
two touch points are horizontal to each other (within a certain
tolerance), then the camera tilt can be changed by dragging the
points up and down. If the touch points are outside that
tolerance, then the rotating the fingers (like spinning a dial)
will rotate the camera around the Y axis.

This is similar to how touch controls work in Apple Maps.