Extracting controller+service for camera controls
Attempting to make the 'Google/Apple Maps' view code less of
a soup. A further step in a future commit is to better define
the MapCamera interface, so that the controller doesn't need
to do calculations to set things like zoom. Instead, it should
be able to blindly pass the UI control value to MapCamera and
let it figure out the rest.