MapCamera calculates zoom/tilt angle, not controller
The controller-type code now only needs to be concerned with
reading/writing the "percentage of maximum" value, and MapCamera
will handle calculating the actual final zoom/tile angle. The
benefit is that the controller logic is simpler by being freed
of this logic, and is more closely focused on handling UI
concerns.