Rotation and zoom are mutually exclusive for touch navigation
That is, you can be rotating the world, or changing the zoom
amount, but not both at the same time.

This makes the touch controls easier to use, by avoiding
inadvertantly zooming the map while rotating your fingers,
or vice-versa. It's difficult/impossible to keep your fingers
exactly the same distance between each other while rotating,
so zooming is unavoidable unless compensated for.

If the Y-rotation delta above a certain threshold, then Y
rotation will be activated, otherwise zoom will be activated.
Tilt rotation still has the highest precedence, and will be
activated no matter what if the fingers are horizontally level.