Camera rotation angle clamped to range from -PI to PI
This ensures the UI slider for the rotation angle (which has
a range of -PI to PI) always has a valid value regardless of
the scene rotation. Otherwise, continually rotating the world
in the same direction using the touch rotation gesture will
cause the slider to max out and get "stuck".