Initial ability to toggle "interactive" and "flythrough" modes
This is a first pass. It would be better if the flythrough
started at your current location in "interactive" mode, and if
stopping the flythrough kept you at the final location. Also,
the interactive navigation controls should be hidden while the
flythrough is running. These things can happen in future commits.