Interactive mode picks up at same place flythrough stopped
Now when stopping the flythrough, the interactive mode starts
at the same place, instead of jumping back to where the
interactive mode was before starting the flythrough. This allows
for a more seamless experience - after clicking "Stop Tour",
the flythrough just stops and you can manually pick up from
where it stopped.

When the camera is perfectly level in flythrough mode, there
will be a small 'jump' when switching back to interactive mode.
This is caused by switching the tilt angle from 0 degrees to
-0.1 degrees (the closest tilt angle allowed by interactive mode,
since it doesn't allow a tilt angle of 0 degrees). This should
be fixed somehow in a future commit.