Parameterizing several flythrough starting position/rotation fields
This is working toward the goal of allowing a flythrough to start
seamlessly from the position/rotation of the interactive camera.
To do this, the flythrough needs to be able to start from any
arbitrary position/rotation, instead of using hard-coded values.

This change doesn't affect the behavior of the default 'flythrough
only' page (because the same previously hard-coded values are
passed in).

The switch from interactive mode to flythrough mode will partially
sync from the interactive camera's starting position/rotation,
but it still has some problems and is not fully seamless. This is
because some fields, such as starting Z position, and not
parameterized yet.