Camera Y-rotates to target point at start of flythrough
Prevents the effect of riding facing backwards in a car, if
the direction towards the target point is opposite of the
interactive camera direction.

Sometimes the rotations are too long, when the target rotation
could be achieved quicker by rotating in the opposite direction.
This should be fixed in the future.