X-rotation more gradual in switch from interactive to flythrough
Before, it very suddenly jerked from the original X rotation to
the target rotation, which was disconcerting and could result in
most of the scene being hidden if the camera was initially looking
straight down.

Now the camera will gradually rotate in the X direction, and reach
the final rotation when the camera reaches the target point. It
looks pretty cool!