Initial camera swoop works if starting from full overhead tilt
Previously, if you were looking straight down at the center of
the city from high up, the camera would suddenly jump to ground
level when starting a flythrough. This could happen if you loaded
a fresh page, set the tilt angle all the way overhead, and
clicked on "Start a Tour".

The reason was that the initial swoop dowm was only accounting
for the X and Z coordinates. Since the X and Z coordinates were
already at the city center, the flythrough logic caused the Y
coordinate to suddenly jump to the target Y coordinate in one
frame.

Now, the swoop logic accounts for all 3 dimensions. The initial
swoop now follows the 3D line between the starting point and
the target point. (I.e. the hypotenuse of the triangle between
the start, target, and ground). So, if the camera is tilted high
above the city from the center, it will slowly move toward the
ground while staying in the same X/Z coordinate.

This isn't perfect though. For some reason, when the camera starts
fully tilted overhead, the flythrough camera sometimes reaches
the target X/Y coordinates before the Z coordinate is reached.
Also, there is a weird extra Y rotation when starting from a
fully overhead angle. But, it is better than it was before.