Flythrough avoids extra Y rotation when camera pointed straight down
As noted in 8ef830ef2e3a24920a762af0ec3e806edea3a9b0, the camera
previously performed a 180deg rotation when starting a flythrough
with the default navigation settings, except for the tilt pointing
straight down. The cause appears to have been floating point issues
with the Z position being close to 0, but not exactly 0.

Now, the max tilt angle is derived, instead of using a hard-coded
magic number. Besides being a little more clear, it seems to fix
the issue by giving the tilt angle a little more precision so that
it is essentially treated like 0. (At least, that's my theory).

It seems like this might have also fixed the issue when the camera
"stalled" in place for a little bit of time after reaching the
center point on the initial swoop.