Using early return to make logic more clear
It took me a minute to figure out what this code did. By changing
it to directly return 1.0 early if within the 'safe from decay'
distance, it reads more clearly to me than the previous approach
of setting a magic value that will be used in a subsequent complex
calculation.