Adding ability to bring all low laying terrain up to minimum height
Produces some interesting effects when the jitter and decay amounts
produce very spiky terrain, by resulting in flat valleys in
between mountains. Sometimes it looks surprisingly good.

However, the details of how to use this need to be worked out more
before enabled it full time, so for now this is not actually being
used in the live code.