Logic to check for nearby bridges is more specific
The logic now tests to see if the neighboring road intersection
is on a bridge, rather than checking to see if it is over water.
While any road intersection over water should also be on a bridge,
checking to see if the road is a bridge is more direct for the
intended purpose of this function (i.e. to prevent bridges from
being too close to each other).