CityEditorController no longer responsible for scene generation
Instead, it sends a message that scene generation should be started
(due to the "Create New City" button being clicked), and an external
message listener is responsible for actually generating the new city.

The benefit of doing this is that CityEditorController is now more
specifically a controller for the "city options" menu, and doesn't
need to be concerned with city generation logic. Ultimately this is
a step toward fixng the memory leak that occurs when creating a new
city, by isolating the setup/teardown code to a central place.