Performing teardown+setup each time a new city is generated
Instead of re-using the same models/controllers for the lifetime
of the page. The reason for doing this is to help isolate and fix
a memory leak that occurs each time a new city is generated.
Destorying and re-creating the relevant object instances should
theoretically prevent the leak from happening. This commit doesn't
fix the leak but it should be a step in that direction.