Destroying scene objects when generating new city
To help prevent a large memory leak when creating a new city.
Apparently, the objects in a scene have to be manually disposed of
when a scene is destroyed to prevent a memory leak. I'm not sure
that this commit 100% fixes the leak, but it seems to allow at least
some garbage collection to occur, according to the Chrome task manager.