Scene is more fully disposed when old city is destroyed
Previously, not all elements in the scene were removed, due to the
underlaying array being changed while it was iterated over. Iterating
backwards fixes that bug and allows all elements in the array to be
disposed.

I'm not sure this 100% fixes the memory leak, but it at least appears
to cause it to happen more slowly, according to the Chrome task manager.