Interactive camera prevented from going below terrain surface
For example, if zoomed all the way in near a hill in hilly terrain.
Previously, there was nothing to prevent the camera from moving under
the terrain, showing the city from underneath. Now, the camera will be
clamped to the terrain surface, regardless of the zoom level.

Not 100% sure that there aren't edge cases with very steep terrain
where you can still see "under" the terrain, but this seems to work for
typical terrain elevations.