First work toward implementing hydraulic erosion of terrain
"Rainfall" adds water to the terrain, and then the water moves down
to lower ground. Evaporation then removes some of the water. The end
result is lakes of water in low laying terring.

In the future, soil can be transported when water flows downhill, causing
the terrain to erode. That does not happen in this commit though.

This implementation is somewhat slow, so it would be nice to figure out
how to make it faster in the future.