Fixing disconnected road networks separate from the main network
The bug was that only west->east or north->south bridges would be
properly built. If a east->west or south->north bridge was attempted
to be built, the bridge would not actually be added to the road network,
but the "continue building road network from the terminal point" _would_
continue. This would result in a separate road network that was not
connected to the main network by a bridge.

As a result of fixing this, there are now more bridges being built than
before. It seems like the "don't build too many bridges in close proximity"
logic also has a bug, which should be fixed in a different commit.