Better algorithm for determining the city center
I.e., the central point that the road network is built out from.
It will now search out from the center for a land tile in ever expanding
squares, instead of searching in a straight line from the center of the
terrain to the edge.

This makes it more likely that a land tile will be found if the terrain
contains a lot of water. It's not perfect though, because if the entire
terrain is covered with water it will result in an error.