Internal terrain data uses 0-n indices, not centered-at-0 indices
This seems like it should make the terrain data easier to work with,
and also be a step toward allowing a mutable Terrain object, which
will be helpful for terrain generation code.

From the outside world, requests for terrain data will still use
centered-at-0 coordinates; only the the interal data structure is being
changed. Coordinates will be translated accordingly internally.