Experimentally shading gradient between land/water terrain
Instead of each triangle being a solid color. A benefit of this
is that single points of water in the terrain are always visible.
A drawback is that when water is next to a steep cliff, the water
appears to "climb" the cliff, which doesn't make visual sense.
So, as is this is probably not the ideal solution for this.