City center can be near terrain edge without blowing up with error
Various bounds are now checked to avoid attempting to generate/process
data outside the bounds of the terrain. Also, the road network bounds
are now set correctly when the road network is not centered near the
center of the terrain.

Currently the city center can't be manually set in a way such that it
wouldn't be near the terrain center, but this should make generation
more robust.