Using strict equality, and being more explicit about undefined vs. null
Not sure why this was checking for null instead of undefined before, but
my hunch is at the time this code was originally written I was unclear on
the difference between the two and so was using them quasi-interchangably.

This commit doesn't directly fix a particular known issue, but is an attempt
to clean up the `RoadNetwork` code a bit in advance of implementing the ability
to remove edges/intersections from an existing `RoadNetwork`.