RoadNetwork.edgeBetween() returns undefined if edge doesn't exist, not false
`undefined` seems like a more idiomatic value for "does not exist".

This change necessitates changing `RoadGeometryBuilder` because it was previously
testing to see if the return value for `edgeBetween()` was `false`. However, rather
than change the code to check for `undefined`, switching the code to use `hasEdgeBetween()`
instead because it is more direct. Finally, also changing some conditions to be more
explicit about the expected value.