PathFinder no longer blows up if a path can't be found
Instead, it returns `undefined` so that the caller can handle that
however it makes sense.

The reason for adding this is to support an upcoming ability to remove
unnecessary roads from a road network. When deciding whether to remove
a row, it will be necessary to know if removing the road results in part
of the road network being cut off.

Besides this specific reason, this makes `PathFinder` more robust in the
face of potential error conditions.