Initial implementation of removing unnecessary roads from road network
That is, roads that aren't near any buildings, and that if removed won't
cut off part of the road network from the rest of the network.

This makes a noticeable improvement on certain cities, especially cities
that have hilly terrain suitable for roads but not buildings - previously
these cities could have a redundant grid of "roads to nowhere" on the
hillsides that would never be built in real life.