Road must have buildings actually touching it to survive culling process
Previously, the only requirement was that a block adjacent to it had a
building in it somewhere. While a simpler thing to check, that could result
in extra roads that didn't seem like they would be built in real life.

A drawback is that this can result in a "maze" of streets on the edge of a
large city that are near each other but can't be reached without traveling
back to a far away intersection. This is something to look into for a future
improvement.