Empty/non-existent blocks always represented by `[]`, not `undefined`
This means that there is a consistent value used when a block has
no buildings (rather than two possible values), which allows code
that uses the `Buildings` data structure to be simpler.

`[]` seems more semantic in this context than `undefined` (i.e.,
`Buildings.blockAtCoordinates` returns a list of buildings, and the
list could be empty).