Fixing edge case with PoleCamera -> InteractiveCamera conversion
When the `PoleCamera` tilt angle is 0 and close to the ground (i.e.
y position is near 0), the zoom factor was previously given an unexpected
value due to the zoom distance being less than the allowed minimum distance.
Now it is set to "fully zoomed in", and the center point is set accordingly
from the zoom distance. This seems like more expected behavior?

It seems like there are still edge cases where the conversion can go off the
rails, so this is probably not the final step toward fully fixing conversion
issues.