Centralizing where 'next target coordinates' are set
With this commit, `determineNextTargetPoint()` can now be called to
set all of the target position/rotation parameters, instead of only
some of them. This seems to make flow of the code easier to understand:
when you want to switch into a different mode (because previous target
was reached, user request, etc.) you can now set the mode (driving, hovering,
bird's eye, etc.) appropriately, and then simply call `determineNextTargetPoint()`.
Previously the logic to set the various parameter targets was split between
`determineNextTargetPoint()` and `tick()`.

There is probably some additional refactoring that can happen in this code
to make the control flow even more clear, but this is a good step.