Initial descent logic for flythrough less of special case in the code
The logic to set the various parameters for the initial descent that
happens at the beginning of a flythrough is now at the same place in the
code as the code to set the parameters for driving/hovering/bird's eye mode.
Previously, these parameters were set in a block of variable declarations
at the top of this function. Besides making the control flow more uniform,
this allows removing several function global variables.

There is more that can be done to clean up/standardize the flow of setting
parameters when a new flying mode/target point is chosen, but this is a good
step in that direction.