Removing unused fields from function arguments
Only the target scene X/Z position is used, so no need to pass in the
other position/rotation fields.

More generally, `VehicleController` should abstract away the details of
which position/rotation fields to set for the current navigation mode. So
in the future it seems like it would be a good idea to avoid the need for
the client to manually set the initial target scene X/Z as well. This way
the client can indicate which high level strategy to navigate (if desired),
but not worry about low level details of how to perform this navigation.