Switch between PoleCamera/InteractiveCamera animates smoothly
Previously, there was a sudden jump at the end of a flythrough if the
position of the PoleCamera and the new position of the InteractiveCamera
differed enough. This could happen because the flythrough allows a wider
range of camera angles than allowed by InteractiveCamera.

Now, the switch between the two cameras will be smoothly animated, which is
much less disorienting.

The control flow of this implementation is a little convoluted, but that
is probably fixable in a future commit.