Single WebGL renderer used across all cities; prevents memory leak
Previously, a new WebGL renderer would be created each time a new
city was created. A memory leak meant that that old renderer was still
around, and eventually would cause the page to either run out of memory,
or the browser to hit it's max number of WebGL instances and start
recycling them.

The only thing that really needs to be disposed between each city is the
ThreeJS scene, while the renderer/canvas element can be re-used for the
lifetime of the page. This change prevents the memory leak, and also has
the benefit of making the code a little easier to understand in my opinion.