Simpler timer/render loop
There is now a single render loop, which handles rendering changes due to
the interactive camera changing position, and the flythrough animation. Scene
rendering will now always happen every 1/60th second, using whatever the most
recent camera position is.

Changing the interactive camera position will not trigger an immediate re-render;
that will happen on the next iteration of the timer loop. In practice there
shouldn't be a need to render any faster than 60fps, and this makes the code
simpler and avoids excess rendering if mouse/touch input triggers many events.

A drawback of this implementation as-is is that the scene will be re-rendered
once every 1/60th second, regardless of whether the camera position has actually
moved or not. In theory it should be possible to optimize this so that the scene
is only re-rendered when the camera position changes, but this can happen in a
future commit.