Single `SceneView` re-used for lifetime of page
Instead of recreating the `SceneView` from scratch each time a new
city is created, only the terrain/road network/buildings data and
scene geometry is recreated and replaced inside the existing `SceneView`
instance. This seems to make the logic simpler.

As part of this, the scene builder logic is broken up into separate
functions that can generate constituent parts of the scene, instead of
a single function that generates the entire scene in one step. This allows
new portions of the scene to be generated without replacing the entire scene.