Removing unused 'fly to bird's eye at city center' animation
Was only used by the previous non-interactive version of the app.
It allowed viewing a bird's eye view of the city, which was useful
because otherwise it was only possible to view the city via a flythrough.
Now that the normal interactive mode allows seeing the city from any
position, and the old version of the app has been removed, there's no
reason to keep this around any longer.