Adding Math.randomInteger(), returns random integer in given range
There are a number of places in the existing code where this function
can be used to make the existing code a little more succinct.

This wraps the implementation provided by ThreeJS, in order to avoid
direct dependencies on ThreeJS outside of the main 3D modeling/rendering
code. If necessary, this can always be replaced with a direct implementation
instead of relying on ThreeJS.