Using correct road network bounds for AerialNavigator
The previous implementation assumed that the city was centered at 0,0
and that the road network extended to the max possible size. This could
cause artifacts with the chosen flight path if this wasn't the case.
(Worst case could be the navigator couldn't find a valid path on a city
centered far away from the terrain center).