AerialNavigator should always terminate, not have infinite loops
No matter what, `AerialNavigator.determineNextTarget()` should always
terminate and result in a new target destination, regardless of the
road network given (even if the road network is completely empty).

First, the candidate axis will be checked for a target. If one can't
be found, the other axis will be checked. Finally if neither axis has
a suitable target, the current position will be chosen as the new target.

Despite being more verbose than the previous implementation, the new
code seems to me to be more clear.