Adding time-based motion generators
This type of motion generator will interpolate between the start and
target values using a given easing function, and be guaranteed to reach
the target after the given number of frames.

In contrast, the previous type of motion generator (such as ClampedLinearMotionGenerator)
apply a certain delta to the initial start value until the target value
is reached.

Time-based interpolation has the advantage of eliminating issues with the
target value not being reached due to floating point precision issues, and
non-linear interpolation methods being easier to reason about in certain
instances.

A drawback is that sometimes "time to target" is not the simplest
way to think about an animation - instead, it can be simpler to think
in terms of "go from here to there at this rate, until you get there".

There doesn't necessarily have to be a single answer for "time-based" vs.
"delta-until-target" based; it's possible having the option to use either
one depending on the circumstance will be the most useful.

As an practical experiment to see how well this approach works, changing
VehicleController in this commit to use the new time-based motion generator.