Adding easing function based on a sine wave
It allows using any portion of a sine wave. For example, this makes
it possible to not only interpolate from the start to the target,
but from the start to the target and then back to the start.

Using the new sine easing function in the 'sync flythrough to interactive
camera' animation as a test.